reorganized care package tests
