+++ LBH.AdultSocialCare.Api.Tests/V1/E2ETests/CarePackages/CarePackagesControllerE2ETests.cs
@@ -1,23 +1,22 @@
-using FluentAssertions;
-using LBH.AdultSocialCare.Api.Tests.V1.Constants;
-using LBH.AdultSocialCare.Api.Tests.V1.DataGenerators;
-using LBH.AdultSocialCare.Api.Tests.V1.Helper;
-using LBH.AdultSocialCare.Api.V1.AppConstants.Enums;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
+using FluentAssertions;
 using HttpServices.Models.Responses;
+using LBH.AdultSocialCare.Api.Tests.V1.Constants;
+using LBH.AdultSocialCare.Api.Tests.V1.DataGenerators;
+using LBH.AdultSocialCare.Api.Tests.V1.Helper;
+using LBH.AdultSocialCare.Api.V1.AppConstants.Enums;
 using LBH.AdultSocialCare.Api.V1.Boundary.CarePackages.Request;
 using LBH.AdultSocialCare.Api.V1.Boundary.CarePackages.Response;
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.CarePackages;
-using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Moq;
 using Xunit;
 
-namespace LBH.AdultSocialCare.Api.Tests.V1.E2ETests.Common
+namespace LBH.AdultSocialCare.Api.Tests.V1.E2ETests.CarePackages
 {
     public class CarePackagesControllerE2ETests : IClassFixture<MockWebApplicationFactory>
     {
