--- LBH.AdultSocialCare.Api.Tests/V1/Helper/DatabaseEntityHelper.cs
@@ -1,24 +1,0 @@
-namespace LBH.AdultSocialCare.Api.Tests.V1.Helper
-{
-
-    public static class DatabaseEntityHelper
-    {
-
-        //public static DatabaseEntity CreateDatabaseEntity()
-        //{
-        //    var entity = new Fixture().Create<Entity>();
-
-        //    return CreateDatabaseEntityFrom(entity);
-        //}
-
-        //public static DatabaseEntity CreateDatabaseEntityFrom(Entity entity)
-        //{
-        //    return new DatabaseEntity
-        //    {
-        //        Id = entity.Id, CreatedAt = entity.CreatedAt
-        //    };
-        //}
-
-    }
-
-}
