+++ LBH.AdultSocialCare.Api.Tests/V1/UseCase/CarePackages/SubmitCarePackageUseCaseTest.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Threading.Tasks;
 using Common.Exceptions.CustomExceptions;
 using FluentAssertions;
 using LBH.AdultSocialCare.Api.Tests.V1.Constants;
@@ -10,11 +12,9 @@ using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.CarePackages;
 using LBH.AdultSocialCare.Api.V1.UseCase.CarePackages.Concrete;
 using Microsoft.AspNetCore.Http;
 using Moq;
-using System;
-using System.Threading.Tasks;
 using Xunit;
 
-namespace LBH.AdultSocialCare.Api.Tests.V1.UseCase.Common
+namespace LBH.AdultSocialCare.Api.Tests.V1.UseCase.CarePackages
 {
     public class SubmitCarePackageUseCaseTest : BaseTest
     {
