+++ LBH.AdultSocialCare.Api.Tests/V1/UseCase/CarePackages/UpdateCarePackageUseCaseTests.cs
@@ -1,18 +1,16 @@
+using System;
+using System.Threading.Tasks;
 using Common.Exceptions.CustomExceptions;
 using FluentAssertions;
 using LBH.AdultSocialCare.Api.Tests.V1.Helper;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Gateways;
-using LBH.AdultSocialCare.Api.V1.Gateways.Common.Interfaces;
-using LBH.AdultSocialCare.Api.V1.UseCase.Common.Concrete;
-using Moq;
-using System;
-using System.Threading.Tasks;
 using LBH.AdultSocialCare.Api.V1.Gateways.CarePackages.Interfaces;
 using LBH.AdultSocialCare.Api.V1.UseCase.CarePackages.Concrete;
+using Moq;
 using Xunit;
 
-namespace LBH.AdultSocialCare.Api.Tests.V1.UseCase.Common
+namespace LBH.AdultSocialCare.Api.Tests.V1.UseCase.CarePackages
 {
     public class UpdateCarePackageUseCaseTests : BaseTest
     {
