+++ LBH.AdultSocialCare.Api.Tests/V1/UseCase/CarePackages/UpsertCarePackageBrokerageUseCaseTest.cs
@@ -1,9 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using Common.Exceptions.CustomExceptions;
 using FluentAssertions;
 using LBH.AdultSocialCare.Api.Tests.Extensions;
 using LBH.AdultSocialCare.Api.Tests.V1.Helper;
 using LBH.AdultSocialCare.Api.V1.AppConstants.Enums;
 using LBH.AdultSocialCare.Api.V1.Domain.CarePackages;
+using LBH.AdultSocialCare.Api.V1.Extensions;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Gateways;
 using LBH.AdultSocialCare.Api.V1.Gateways.CarePackages.Interfaces;
@@ -12,13 +16,9 @@ using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.CarePackages;
 using LBH.AdultSocialCare.Api.V1.UseCase.CarePackages.Concrete;
 using Microsoft.AspNetCore.Http;
 using Moq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using LBH.AdultSocialCare.Api.V1.Extensions;
 using Xunit;
 
-namespace LBH.AdultSocialCare.Api.Tests.V1.UseCase.Common
+namespace LBH.AdultSocialCare.Api.Tests.V1.UseCase.CarePackages
 {
     public class UpsertCarePackageBrokerageUseCaseTest : BaseTest
     {
